+++ tests/TestApp/application/models/Comment.php
@@ -2,6 +2,8 @@
 
 class Comment extends Rougin\Wildfire\CodeigniterModel
 {
+    use Rougin\Wildfire\Traits\ValidateTrait;
+
     /**
      * Columns that will be hidden in the display.
      * If not set, it will hide a "password" column if it exists.
@@ -9,4 +11,14 @@ class Comment extends Rougin\Wildfire\CodeigniterModel
      * @var array
      */
     protected $hidden = array('id');
+
+    /**
+     * An array of validation rules. This needs to be the same format
+     * as validation rules passed to the Form_validation library.
+     *
+     * @var array
+     */
+    protected $validation_rules = array(
+        array('field' => 'name', 'label' => 'Name', 'rules' => 'required'),
+    );
 }